# openapi/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Energy Allocation API
  version: 1.0.0
  description: >-
    Tenant-aware API for onboarding facilities, ingesting data, generating model-driven
    recommendations, and tracking realized savings.
servers:
  - url: https://api.energyallocation.com
security:
  - oauth2: [client_admin, facility_manager, energy_analyst]
paths:
...
  /v1/integrations/partners:
    post:
      tags: [Partner Integrations]
      summary: Register a new third-party partner
      operationId: registerPartner
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PartnerCreate' }
      responses:
        '201':
          description: Partner registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Partner' }
    get:
      tags: [Partner Integrations]
      summary: List all registered partners
      operationId: listPartners
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [active, suspended, archived] }
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Partner list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Partner' }
                  total: { type: integer }
  /v1/integrations/partners/{partner_id}:
    get:
      tags: [Partner Integrations]
      summary: Get partner details
      operationId: getPartner
      parameters:
        - in: path
          name: partner_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Partner details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Partner' }
    patch:
      tags: [Partner Integrations]
      summary: Update partner metadata
      operationId: updatePartner
      parameters:
        - in: path
          name: partner_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PartnerUpdate' }
      responses:
        '200':
          description: Partner updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Partner' }
  /v1/integrations/partners/{partner_id}/keys:
    post:
      tags: [Partner API Keys]
      summary: Create a new API key for a partner
      operationId: createPartnerApiKey
      parameters:
        - in: path
          name: partner_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApiKeyCreate' }
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKeyWithSecret' }
    get:
      tags: [Partner API Keys]
      summary: List API keys for a partner
      operationId: listPartnerApiKeys
      parameters:
        - in: path
          name: partner_id
          required: true
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [active, revoked] }
      responses:
        '200':
          description: API keys list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ApiKey' }
                  partner_id: { type: string }
  /v1/integrations/partners/{partner_id}/keys/{key_id}:
    patch:
      tags: [Partner API Keys]
      summary: Rotate or revoke an API key
      operationId: rotatePartnerApiKey
      parameters:
        - in: path
          name: partner_id
          required: true
          schema: { type: string }
        - in: path
          name: key_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [rotate, revoke]
              required: [action]
      responses:
        '200':
          description: API key updated
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/ApiKeyWithSecret' }
                  - type: object
                    properties:
                      key_id: { type: string }
                      status: { type: string, enum: [revoked] }
  /v1/integrations/partners/{partner_id}/webhooks:
    post:
      tags: [Partner Webhooks]
      summary: Register a webhook endpoint
      operationId: createPartnerWebhook
      parameters:
        - in: path
          name: partner_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WebhookCreate' }
      responses:
        '201':
          description: Webhook created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Webhook' }
    get:
      tags: [Partner Webhooks]
      summary: List webhooks for a partner
      operationId: listPartnerWebhooks
      parameters:
        - in: path
          name: partner_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Webhooks list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Webhook' }
                  partner_id: { type: string }
  /v1/integrations/partners/{partner_id}/webhooks/{webhook_id}:
    patch:
      tags: [Partner Webhooks]
      summary: Update a webhook
      operationId: updatePartnerWebhook
      parameters:
        - in: path
          name: partner_id
          required: true
          schema: { type: string }
        - in: path
          name: webhook_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WebhookUpdate' }
      responses:
        '200':
          description: Webhook updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Webhook' }
  /v1/integrations/partners/{partner_id}/allocations:
    post:
      tags: [Partner Allocations]
      summary: Assign partner to a client
      operationId: assignPartnerToClient
      parameters:
        - in: path
          name: partner_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AllocationCreate' }
      responses:
        '201':
          description: Allocation created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Allocation' }
    get:
      tags: [Partner Allocations]
      summary: List partner allocations
      operationId: listPartnerAllocations
      parameters:
        - in: path
          name: partner_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Allocations list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Allocation' }
                  partner_id: { type: string }
  /v1/integrations/partners/{partner_id}/allocations/{allocation_id}:
    patch:
      tags: [Partner Allocations]
      summary: Update or revoke an allocation
      operationId: updatePartnerAllocation
      parameters:
        - in: path
          name: partner_id
          required: true
          schema: { type: string }
        - in: path
          name: allocation_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AllocationUpdate' }
      responses:
        '200':
          description: Allocation updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Allocation' }
  /v1/integrations/partners/{partner_id}/usage:
    get:
      tags: [Partner Usage]
      summary: Get usage statistics and audit logs
      operationId: getPartnerUsage
      parameters:
        - in: path
          name: partner_id
          required: true
          schema: { type: string }
        - in: query
          name: start_date
          schema: { type: string, format: date }
        - in: query
          name: end_date
          schema: { type: string, format: date }
      responses:
        '200':
          description: Usage and audit data
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PartnerUsage' }
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://auth.energyallocation.com/oauth/token
          scopes:
            client_admin: Manage client-level resources
            facility_manager: Manage facility operations
            energy_analyst: Read analytics and model outputs
            viewer: Read-only access
            ops_admin: Internal operations access
            ml_engineer: Internal ML administration
  parameters:
...
    limit: { in: query, name: limit, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
...
  schemas:
...
    PartnerCreate:
      type: object
      properties:
        name: { type: string }
        industry: { type: string }
        contact_email: { type: string, format: email }
        contact_name: { type: string }
        scopes:
          type: array
          items: { type: string }
        assigned_clients:
          type: array
          items: { type: string }
      required: [name, contact_email, contact_name]
    Partner:
      type: object
      properties:
        partner_id: { type: string }
        name: { type: string }
        industry: { type: string }
        contact_email: { type: string, format: email }
        contact_name: { type: string }
        status: { type: string, enum: [active, suspended, archived] }
        assigned_clients:
          type: array
          items: { type: string }
        assigned_scopes:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
      required: [partner_id, name, status, created_at]
    PartnerUpdate:
      type: object
      properties:
        status: { type: string, enum: [active, suspended, archived] }
        assigned_clients:
          type: array
          items: { type: string }
        assigned_scopes:
          type: array
          items: { type: string }
    ApiKeyCreate:
      type: object
      properties:
        name: { type: string }
        scopes:
          type: array
          items: { type: string }
      required: [name]
    ApiKey:
      type: object
      properties:
        key_id: { type: string }
        partner_id: { type: string }
        name: { type: string }
        scopes:
          type: array
          items: { type: string }
        status: { type: string, enum: [active, revoked] }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        last_used_at: { type: [string, 'null'], format: date-time }
      required: [key_id, partner_id, name, status, created_at, expires_at]
    ApiKeyWithSecret:
      type: object
      properties:
        key_id: { type: string }
        secret: { type: string }
        partner_id: { type: string }
        scopes:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        warning: { type: string }
      required: [key_id, secret, partner_id, created_at, expires_at]
    WebhookCreate:
      type: object
      properties:
        url: { type: string, format: uri }
        description: { type: string }
        events:
          type: array
          items: { type: string }
        max_retries: { type: integer, default: 3 }
        backoff_seconds: { type: integer, default: 60 }
      required: [url, events]
    WebhookUpdate:
      type: object
      properties:
        url: { type: string, format: uri }
        events:
          type: array
          items: { type: string }
        status: { type: string, enum: [active, paused] }
    Webhook:
      type: object
      properties:
        webhook_id: { type: string }
        partner_id: { type: string }
        url: { type: string, format: uri }
        description: { type: string }
        status: { type: string, enum: [active, paused] }
        events:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
        signature_algorithm: { type: string }
      required: [webhook_id, partner_id, url, status, events, created_at]
    AllocationCreate:
      type: object
      properties:
        client_id: { type: string }
        scopes:
          type: array
          items: { type: string }
        read_only: { type: boolean, default: true }
      required: [client_id]
    AllocationUpdate:
      type: object
      properties:
        status: { type: string, enum: [active, revoked] }
        scopes:
          type: array
          items: { type: string }
        read_only: { type: boolean }
    Allocation:
      type: object
      properties:
        allocation_id: { type: string }
        partner_id: { type: string }
        client_id: { type: string }
        status: { type: string, enum: [active, revoked] }
        scopes:
          type: array
          items: { type: string }
        read_only: { type: boolean }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
      required: [allocation_id, partner_id, client_id, status, scopes, created_at]
    PartnerUsage:
      type: object
      properties:
        partner_id: { type: string }
        summary:
          type: object
          properties:
            total_events: { type: integer }
            api_keys: { type: integer }
            allocations: { type: integer }
            webhooks: { type: integer }
        audit_events:
          type: array
          items:
            type: object
            properties:
              audit_id: { type: string }
              action: { type: string }
              resource_id: { type: string }
              actor: { type: string }
              timestamp: { type: string, format: date-time }
              metadata:
                type: object
                additionalProperties: true
      required: [partner_id, summary, audit_events]